refactored Webserver to remove boolean arguments from constructor and split into two types
